requirements.txt
django==1.8.5
djangorestframework==3.3.2
django-rest-swagger==0.3.4